feat move folder to pages to organice folder structure, add sidebar and dashboard styles
