+++ angular.json
@@ -24,9 +24,7 @@
               "src/assets",
               {
                 "glob": "**/*",
-                "input": "public",
-                 "output": "/assets/"
-                
+                "input": "public"
               }
             ],
             "styles": [
