+++ src/app/app.routes.ts
@@ -1,14 +1,14 @@
 import { Routes } from '@angular/router';
 import { LoginComponent } from './auth/login/login.component';
 import { RegisterComponent } from './auth/register/register.component';
-import { ClientsComponent } from './dashboard/clients/clients.component';
-import { AppointmentsComponent } from './dashboard/appointments/appointments.component';
-import { FinancialComponent } from './dashboard/financial/financial.component';
-import { InventoryComponent } from './dashboard/inventory/inventory.component';
-import { ServicesComponent } from './dashboard/services/services.component';
+import { ClientsComponent } from './pages/clients/clients.component';
+import { AppointmentsComponent } from './pages/appointments/appointments.component';
+import { FinancialComponent } from './pages/financial/financial.component';
+import { InventoryComponent } from './pages/inventory/inventory.component';
+import { ServicesComponent } from './pages/services/services.component';
 import { NotfoundComponent } from './pages/notfound/notfound.component';
 import { AuthGuard } from './auth/guards/auth.guard';
-import { DashboardComponent } from './dashboard/dashboard/dashboard.component';
+import { DashboardComponent } from './pages/dashboard/dashboard.component';
 
 export const routes: Routes = [
     { path: '', redirectTo: '/login', pathMatch: 'full' },
