--- src/app/dashboard/clients/clients.component.html
@@ -1,1 +1,0 @@
-<p>clients works!</p>
