--- src/app/dashboard/dashboard/dashboard.component.html
@@ -1,4 +1,0 @@
-<app-navbar></app-navbar>
-<div class="container">
-  <router-outlet></router-outlet>
-</div>
