+++ src/app/pages/clients/clients.component.html
@@ -1,0 +1,2 @@
+<img src="images/working-on-svg.svg" alt="working on svg">
+
