+++ src/app/pages/dashboard/dashboard.component.html
@@ -1,0 +1,18 @@
+<div class="container">
+  <app-navbar class="header"></app-navbar>
+  <app-sidebar class="sidebar"></app-sidebar>
+  <main class="content">
+    <!-- <div class="cards">
+      <div class="card">📅 Próximas Citas</div>
+      <div class="card">💰 Ingresos Mensuales</div>
+      <div class="card">📊 Estadísticas</div>
+    </div>
+    
+    <div class="charts">
+      <div class="chart">📈 Gráfico de Crecimiento</div>
+      <div class="chart">📉 Gráfico de Ventas</div>
+    </div> -->
+    <router-outlet></router-outlet>
+  </main>
+  <app-footer class="footer"></app-footer>
+</div>
