+++ src/app/pages/dashboard/dashboard.component.ts
@@ -1,10 +1,12 @@
 import { Component } from '@angular/core';
 import { NavbarComponent } from "../../shared/components/navbar/navbar.component";
 import { RouterOutlet } from '@angular/router';
+import { SidebarComponent } from "../../shared/components/sidebar/sidebar.component";
+import { FooterComponent } from "../../shared/components/footer/footer.component";
 
 @Component({
   selector: 'app-dashboard',
-  imports: [RouterOutlet, NavbarComponent],
+  imports: [RouterOutlet, NavbarComponent, SidebarComponent, FooterComponent],
   templateUrl: './dashboard.component.html',
   styleUrl: './dashboard.component.css'
 })
