+++ src/app/pages/inventory/inventory.component.html
