+++ src/app/pages/inventory/inventory.component.spec.ts
