+++ src/app/pages/inventory/inventory.component.ts
