+++ src/app/shared/components/footer/footer.component.css
@@ -1,0 +1,9 @@
+footer {
+    width: 100%;
+    height: 50px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    background: #222;
+    color: white;
+  }
