+++ src/app/shared/components/footer/footer.component.html
@@ -1,1 +1,3 @@
-<p>footer works!</p>
+<footer class="footer">
+    <p>© 2025 Hirify SaaS</p>
+  </footer>
