+++ src/app/shared/components/navbar/navbar.component.css
@@ -1,8 +1,19 @@
-.navbar{
-    padding: .5rem;
-    display: flex;
-    background-color: #E5EDF1;
-    justify-content: space-between;
-    align-content: center;
-    align-items: center;
+.logo {
+  padding: 0.5rem;
+  margin: 0;
+  color: #37474F;
 }
+.navbar {
+  width: 100%;
+  height: 60px;
+  display: flex;
+  background-color: #96C2DB;
+  justify-content: space-between;
+  align-content: center;
+  align-items: center;
+}
+
+.navbar-button {
+    padding: 1rem;
+  }
+
