+++ src/app/shared/components/navbar/navbar.component.html
@@ -1,5 +1,5 @@
 <div class="navbar">
-    <h1>Hirify</h1>
+    <h1 class="logo">Hirify</h1>
     <div class="navbar-button">
         <button mat-button (click)="logOut()">Salir</button>
     </div>
