+++ src/app/shared/components/sidebar/sidebar.component.css
@@ -1,0 +1,34 @@
+aside {
+  background: #e5edf1;
+  padding: 10px 0;
+  height: 100%;
+}
+
+.sidebar nav ul {
+  list-style: none;
+  padding: 0;
+  padding: .5rem;
+}
+
+.sidebar nav ul li {
+  margin: 10px 0;
+  width: 100%;
+}
+
+.sidebar nav ul li a{
+  text-decoration: none;
+  display: block;
+  width: 100%;
+  height: 100%;
+  color: #37474f;
+  padding: 18px;
+  font-size: larger;
+}
+
+.sidebar nav ul li a:hover  {
+  color: #e5edf1;
+}
+.sidebar nav ul li:hover{
+  cursor: pointer;
+  background-color: #547A94;
+}
