+++ src/app/shared/components/sidebar/sidebar.component.ts
@@ -1,4 +1,5 @@
 import { Component } from '@angular/core';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-sidebar',
@@ -7,5 +8,9 @@ import { Component } from '@angular/core';
   styleUrl: './sidebar.component.css'
 })
 export class SidebarComponent {
+  constructor(private router: Router) { }
 
+  navigateToChildrens(route: string) {
+    this.router.navigate([`/dashboard/${route}`])
+  }
 }
